mock sftpservice
