+++ src/backend/libs/efiling-demo-starter/src/test/java/ca/bc/gov/open/jag/efiling/demo/AutoConfigurationTest.java
@@ -7,12 +7,13 @@ import ca.bc.gov.open.jag.efilingcommons.model.EfilingPayment;
 import ca.bc.gov.open.jag.efilingcommons.payment.PaymentAdapter;
 import ca.bc.gov.open.jag.efilingcommons.service.*;
 import ca.bc.gov.open.jag.efilingcommons.submission.EfilingStatusService;
+import ca.bc.gov.open.sftp.starter.SftpService;
 import org.junit.jupiter.api.*;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
-import org.springframework.cache.CacheManager;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.data.redis.serializer.Jackson2JsonRedisSerializer;
 
+import java.io.ByteArrayInputStream;
 import java.math.BigDecimal;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -40,9 +41,18 @@ public class AutoConfigurationTest {
             assertThat(it).hasSingleBean(EfilingSubmissionService.class);
             assertThat(it).hasSingleBean(EfilingCourtLocationService.class);
             assertThat(it).hasSingleBean(EfilingStatusService.class);
-            assertThat(it).hasSingleBean(CacheManager.class);
+            assertThat(it).hasBean("demoAccountCacheManager");
+            assertThat(it).hasBean("demoDocumentCacheManager");
             assertThat(it).hasSingleBean(Jackson2JsonRedisSerializer.class);
             assertThat(it).hasSingleBean(BCeIDAccountService.class);
+        });
+    }
+
+    @Test
+    @DisplayName("OK: test of paymentAdapterMock")
+    public void testMockPaymentAdapter() {
+        context.run(it -> {
+
             assertThat(it).hasSingleBean(PaymentAdapter.class);
 
             PaymentAdapter paymentAdapter = it.getBean(PaymentAdapter.class);
@@ -51,11 +61,29 @@ public class AutoConfigurationTest {
             Assertions.assertEquals("APP", paymentAdapter.makePayment(new EfilingPayment(BigDecimal.TEN, BigDecimal.ONE, "invoice", "client")).getApprovalCd());
             Assertions.assertEquals("DEC", paymentAdapter.makePayment(new EfilingPayment(BigDecimal.TEN, BigDecimal.TEN, "invoice", "client")).getApprovalCd());
 
-            GetAccountRequest request =  GetAccountRequest.BusinessSelfRequest("test");
+            GetAccountRequest request = GetAccountRequest.BusinessSelfRequest("test");
             Assertions.assertEquals("efilehub test account", bCeIDAccountService.getIndividualIdentity(request).get().getName().getFirstName());
+
 
         });
     }
 
+    @Test
+    @DisplayName("OK: test of sftpServiceMock")
+    public void testMockSftpAdapter() {
+        context.run(it -> {
+
+            assertThat(it).hasSingleBean(SftpService.class);
+
+            SftpService sut = it.getBean(SftpService.class);
+
+            Assertions.assertDoesNotThrow(() -> sut.put(new ByteArrayInputStream("test".getBytes()), "location"));
+            Assertions.assertDoesNotThrow(() -> sut.getContent("any"));
+            Assertions.assertDoesNotThrow(() -> sut.listFiles("any"));
+            Assertions.assertDoesNotThrow(() -> sut.moveFile("a", "b"));
+
+
+        });
+    }
 
 }
